Raw files now deleted after successful staging.
The staging bucket is our "source of truth", where data (with the correct tags and metadata) is persisted for as long as we require it. Deleting from raw on both staging success and failure prevents duplication of data, and also means we can verify staging health by the raw bucket either being empty or only containing files that are currently being staged. Using the WaitForRawBucketReadsToComplete step is not advised - it is here to provide backward compatibility with organizations reading data directly from the raw bucket after receiving an S3 trigger. If you need low-latency / near real-time access to new data, consider implementing the lambda architecture with kinesis data streams and kinesis firehose. If you have to read from the raw bucket for certain data sources, consider making this wait state configurable.